Bimp Doc
Former-commit-id: fe29ca499132284157c7ee736646a3feb7b833b7

+++ htdocs/bimpcore/classes/BimpDocumentation.php
@@ -1,4 +1,5 @@
 <?php
+
 class BimpDocumentation {
 
     var $lines = array();
@@ -12,24 +13,40 @@ class BimpDocumentation {
     var $admin = true;
     var $mode = 'link';
     var $idSection = '';
-    
+    var $initNiveau = 0;
+    var $initPrefMenu = '';
+
     function __construct($type, $name, $mode = 'link', $idSection = 'princ', &$menu = array()) {
+        if (!is_array($menu)){
+            $this->initPrefMenu = $menu;
+            $this->initialiseArrayMenu($menu);
+        }
         $this->name = $name;
         $this->type = $type;
         $this->mode = $mode;
         $this->menu = &$menu;
         $this->idSection = $idSection;
         static::$nbInstance++;
-        if(static::$nbInstance > 20){
+        if (static::$nbInstance > 20) {
             BimpCore::addlog('Attention boucle dans DOC');
             die('Probléme technique, contacté l\'équipe dév');
         }
     }
-    
-    
 
-    public static function renderBtn($name, $text = '')
-    {
+    private function initialiseArrayMenu(&$menu) {
+        $tabT = explode('.', $menu);
+        $menu = array();
+        for ($i = 1; $i <= 10; $i++) {
+            if (isset($tabT[$i - 1])) {
+                for ($y = 1; $y <= intval($tabT[$i - 1]); $y++) {
+                    $menu[$i][] = '';
+                }
+                $this->initNiveau = $i - 1;
+            }
+        }
+    }
+
+    public static function renderBtn($name, $text = '') {
         $onClickInit = "docModal.loadAjaxContent($(this), 'loadDocumentation', {name: '" . $name . "'}, 'Doc : " . $name . "', 'Chargement', function (result, bimpAjax) {});";
 
         $onClickInit .= 'docModal.show();';
@@ -45,22 +62,32 @@ class BimpDocumentation {
     }
 
     static function getPathFile($type, $name, $update_file = false) {
-        if($update_file)
-            return DOL_DATA_ROOT . '/bimpcore/docs/'.static::$lang.'/' .($update_file?'upd_' : '') . $name . '_' . $type . '.txt';
+        if ($update_file)
+            return DOL_DATA_ROOT . '/bimpcore/docs/' . static::$lang . '/' . ($update_file ? 'upd_' : '') . $name . '_' . $type . '.txt';
         else
-            return DOL_DOCUMENT_ROOT . '/bimpcore/docs/'.static::$lang.'/' .($update_file?'upd_' : '') . $name . '_' . $type . '.txt';
+            return DOL_DOCUMENT_ROOT . '/bimpcore/docs/' . static::$lang . '/' . ($update_file ? 'upd_' : '') . $name . '_' . $type . '.txt';
     }
 
     function getThisPathFile() {
         $file = static::getPathFile($this->type, $this->name, true);
-        if(is_file($file)){
-            if(preg_replace('~\R~u', "\r\n", file_get_contents($file)) == preg_replace('~\R~u', "\r\n", file_get_contents(static::getPathFile($this->type, $this->name)))){
+        if (is_file($file)) {
+            if (is_file(static::getPathFile($this->type, $this->name)) && preg_replace('~\R~u', "\r\n", file_get_contents($file)) == preg_replace('~\R~u', "\r\n", file_get_contents(static::getPathFile($this->type, $this->name)))) {
                 unlink($file);
-            }
-            else
+            } else
                 return $file;
         }
         return static::getPathFile($this->type, $this->name);
+    }
+
+    private function getPrefMenu($niveau) {
+        $pref = '';
+        for ($i = 1; $i <= 10; $i++) {//10 niveau de titres
+            if ($i <= $niveau)
+                $pref .= (isset($this->menu[$i]) ? count($this->menu[$i]) : 0) . ($i < $niveau ? '.' : '');
+            else
+                unset($this->menu[$i]);
+        }
+        return $pref;
     }
 
     function initLines($initNiveau = 0) {
@@ -73,54 +100,49 @@ class BimpDocumentation {
             if (preg_match('#\{code}([^\[]*) ?#', $ln, $matches)) {
                 $ln = '<xmp>' . $matches[1] . '</xmp>';
             } elseif (preg_match('#([^\[]*)?{{([^\[]*)?}}#U', $ln, $matches) && isset($matches[2])) {
-                $idSection = $this->idSection.'_'.count($this->lines);
-                $this->lines[] = array('div' => array('id'=>$idSection));
-                if(is_file(static::getPathFile('doc', $matches[2])) || is_file(static::getPathFile('doc', $matches[2], true))){
+                $idSection = $this->idSection . '_' . count($this->lines);
+                $this->lines[] = array('div' => array('id' => $idSection));
+                if (is_file(static::getPathFile('doc', $matches[2])) || is_file(static::getPathFile('doc', $matches[2], true))) {
                     $child = new BimpDocumentation('doc', $matches[2], $this->mode, $idSection, $this->menu);
+                    $pref = $this->getPrefMenu($niveau+1);
+//                    echo '<pre>';print_r($this->menu);
+
                     $child->initLines($niveau);
-                    if($this->admin)
-                        $this->lines[] = array('value' => static::btnEditer($matches[2], $idSection));
+                    if ($this->admin)
+                        $this->lines[] = array('value' => static::btnEditer($matches[2], $idSection, $pref));
                     $this->lines = BimpTools::merge_array($this->lines, $child->lines);
                     $this->errors = BimpTools::merge_array($this->errors, $child->errors);
+                } elseif ($this->admin) {
+                    $this->lines[] = array('value' => static::btnEditer($matches[2], $idSection, $pref));
                 }
-                else{
-                    $this->lines[] = array('value' => static::btnEditer($matches[2], $idSection));
-                }
-                $this->lines[] = array('div' => array('close'=>true));
+                $this->lines[] = array('div' => array('close' => true));
                 continue;
             } elseif (preg_match('#([^\[]*)?{{([^\[]*)?}#U', $ln, $matches) && isset($matches[2])) {
                 $datas = explode(':', $matches[2]);
-                
-                $html = '<img src="'.DOL_URL_ROOT.'/bimpcore/docs/img/'.$this->convertTag($datas[0]).'"';
-                if(isset($datas[1]))
-                    $html .= ' width="'.$datas[1].'"';
-                if(isset($datas[2]))
-                    $html .= ' height="'.$datas[2].'"';
+
+                $html = '<img src="' . DOL_URL_ROOT . '/bimpcore/docs/img/' . $this->convertTag($datas[0]) . '"';
+                if (isset($datas[1]))
+                    $html .= ' width="' . $datas[1] . '"';
+                if (isset($datas[2]))
+                    $html .= ' height="' . $datas[2] . '"';
                 $html .= '/>';
-                $ln = str_replace('{{'.$matches[2].'}', $html, $ln);
+                $ln = str_replace('{{' . $matches[2] . '}', $html, $ln);
             } elseif (preg_match('#(={2,4}) ?([^\[]*) ?#', $ln, $matches)) {
                 $ln = $this->traiteLn($matches[2]);
                 $niveau = (strlen($matches[1]) - 1) + $initNiveau;
                 $type = 'h' . $niveau;
-                
+
                 /* gestion menu */
-                if($this->type == 'doc'){
-                    $this->menu[$niveau][] = $ln;
-                    $pref = '';
-                    for($i=1;$i<=10;$i++){//10 niveau de titres
-                        if($i<=$niveau)
-                        $pref .= count($this->menu[$i]).($i<$niveau ? '.' : '');
-                        else
-                            unset($this->menu[$i]);
-                    }
-                    $ln = $pref . ' '.$ln;
-                    $hash = 'menu'.$pref;
-                    if ($niveau > 1){
+                if ($this->type == 'doc') {
+                    $this->menu[$niveau][] = '';
+                    $pref = $this->getPrefMenu($niveau);
+                    $ln = $pref . ' ' . $ln;
+                    $hash = 'menu' . $pref;
+                    if ($niveau > 1) {
                         $styleMenu[] = 'text-indent: ' . ($niveau * 10) . 'px;';
                         $styleCore[] = 'padding-left: ' . ($niveau * 10) . 'px;';
                     }
                 }
-                
             } else {
                 $ln = $this->traiteLn($ln);
             }
@@ -128,18 +150,25 @@ class BimpDocumentation {
             $this->lines[] = array('balise' => $type, 'value' => $ln, 'styleMenu' => $styleMenu, 'styleCore' => $styleCore, 'hash' => $hash);
         }
     }
-    
-    static function convertTag($in){
+
+    static function convertTag($in) {
         $in = strtolower($in);
         $in = str_replace(' ', '_', $in);
         return $in;
     }
-    
-    function getFileData($type = 'text'){
-        if($type == 'text')
-            return file_get_contents($this->getThisPathFile());
-        elseif($type == 'array')
-            return file($this->getThisPathFile(), FILE_IGNORE_NEW_LINES);
+
+    function getFileData($type = 'text') {
+        
+        if ($type == 'text'){
+            if(is_file($this->getThisPathFile()))
+                return file_get_contents($this->getThisPathFile());
+            return '';
+        }
+        elseif ($type == 'array'){
+            if(is_file($this->getThisPathFile()))
+                return file($this->getThisPathFile(), FILE_IGNORE_NEW_LINES);
+            return array();
+        }
     }
 
     function getEditFormDoc() {
@@ -153,43 +182,45 @@ class BimpDocumentation {
     }
 
     function saveFile($type, $name, $value) {
-        if(!file_put_contents($this->getPathFile($type, $name, true), $value))
-                $this->errors[] = 'Enregistrement impossible : '.$this->getPathFile($type, $name, true);
+        if (!file_put_contents($this->getPathFile($type, $name, true), $value))
+            $this->errors[] = 'Enregistrement impossible : ' . $this->getPathFile($type, $name, true);
     }
-    
-    function btnEditer($name, $idSection){
-        if($this->mode == 'modal')
-            $onClick = "editSection('".$idSection."', '".$name."');";
-        else
-            $onClick = 'window.location = \''.static::getLink($name, array('action'=>'edit')).'\'';
-        
-        return '<button onclick="'.$onClick.'">'.BimpRender::renderIcon('fas_edit').'</button>';
+
+    function btnEditer($name, $idSection, $serializedMenu = '[]') {
+        if ($this->mode == 'modal') {
+            $onClick = "editDocumentation('" . $idSection . "', '" . $name . "', '" . addslashes($serializedMenu) . "');";
+        } else
+            $onClick = 'window.location = \'' . static::getLink($name, array('action' => 'edit')) . '\'';
+
+        return '<button onclick="' . $onClick . '">' . BimpRender::renderIcon('fas_edit') . $niveau . '</button>';
     }
-    
-    function getDoc(){
+
+    function getDoc() {
         return $this->getFileData();
     }
 
-    function displayDoc() {
+    function displayDoc($mode = 'text') {
         if (!is_file($this->getThisPathFile())) {
             $this->errors[] = 'Doc ' . $this->name . ' introuvable : ' . $this->getThisPathFile();
             return '';
         }
 
-        $this->initLines();
-        
+        $this->initLines($this->initNiveau);
+
         $sections = array();
-        $sections[] = '';
-        if($this->admin)
-            $sections[] = static::btnEditer($this->name, $this->idSection);
-        
-        $sections[] = $this->getMenu();
-        $sections[] = $this->getCore();
-        
+        $sections['vide'] = '';
+
+        $sections['menu'] = $this->getMenu();
+        $sections['core'] = $this->getCore();
+
 
         //impression
-        $sections = array_filter($sections);
-        return '<div id="'.$this->idSection.'">'.implode('<br/><br/>', $sections).'</div>';
+        if ($mode == 'text') {
+            $sections = array_filter($sections);
+            return implode('<br/><br/>', $sections);
+        } elseif ($mode == 'array') {
+            return array('menu' => $sections['menu'], 'core' => $sections['core']);
+        }
 
 
 //        if ($htmlMenu != '')
@@ -197,26 +228,28 @@ class BimpDocumentation {
 //        else
 //            return $htmlContent;
     }
-    
-    static function getLink($name, $params = array()){
+
+    static function getLink($name, $params = array()) {
         $params['name'] = $name;
         $newParams = array();
-        foreach($params as $clef => $val){
-            $newParams[] = $clef.'='.$val;
+        foreach ($params as $clef => $val) {
+            $newParams[] = $clef . '=' . $val;
         }
-        return DOL_URL_ROOT.'/bimpcore/docs/test.php?'.implode('&', $newParams);
+        return DOL_URL_ROOT . '/bimpcore/docs/test.php?' . implode('&', $newParams);
     }
 
     function displayInfoPop() {
         $this->initLines();
         return $this->getCore();
     }
-    
-    
 
     function getMenu() {
-        $html = '';
+        $html = '<div id="' . $this->idSection . '_menu">';
         foreach ($this->lines as $info) {
+            if (isset($info['div'])) {
+                $html .= '<' . (isset($info['div']['close']) ? '/' : '') . 'div' . (isset($info['div']['id']) ? ' id="' . $info['div']['id'] . '_menu"' : '') . (isset($info['div']['class']) ? ' id="' . $info['div']['class'] . '"' : '') . '>';
+            }
+
             if ($info['balise'] != '') {
                 $htmlContent .= '<' . $info['balise'];
                 if ($info['hash']) {
@@ -224,27 +257,30 @@ class BimpDocumentation {
                 }
             }
         }
+        $html .= '</div>';
         return $html;
     }
 
     function getCore() {
         $this->pOpen = true;
-        $html = '<div>';
+        $html = '<div id="' . $this->idSection . '">';
+        if ($this->admin)
+            $html .= static::btnEditer($this->name, $this->idSection, $this->initPrefMenu) . '<br/>';
         foreach ($this->lines as $info) {
-            if(isset($info['div'])){
-                if($this->pOpen){
+            if (isset($info['div'])) {
+                if ($this->pOpen) {
                     $html .= '</div>';
                     $this->pOpen = false;
                 }
-                $html .= '<'.(isset($info['div']['close'])? '/' : '').'div'.(isset($info['div']['id'])? ' id="'.$info['div']['id'].'"' : '').(isset($info['div']['class'])? ' id="'.$info['div']['class'].'"' : '').'>';
+                $html .= '<' . (isset($info['div']['close']) ? '/' : '') . 'div' . (isset($info['div']['id']) ? ' id="' . $info['div']['id'] . '"' : '') . (isset($info['div']['class']) ? ' id="' . $info['div']['class'] . '"' : '') . '>';
             }
-            
-            
+
+
             if ($info['balise'] != '') {
-                if($this->pOpen){
+                if ($this->pOpen) {
                     $html .= '</div>';
                 }
-                $html .= "<div".(isset($info['styleCore']) && count($info['styleCore']) ? ' style="' . implode(' ', $info['styleCore']) . '"' : '').">";
+                $html .= "<div" . (isset($info['styleCore']) && count($info['styleCore']) ? ' style="' . implode(' ', $info['styleCore']) . '"' : '') . ">";
                 $this->pOpen = true;
                 $html .= '<' . $info['balise'];
                 if ($info['hash']) {
@@ -255,6 +291,7 @@ class BimpDocumentation {
                 $html .= $info['value'] . '<br/>';
         }
         $html .= '</div>';
+//        $html .= '<pre>'.print_r($this->lines,1);
         return $html;
     }
 
@@ -277,8 +314,8 @@ class BimpDocumentation {
 //            die('ttt');
             if (isset($matches[2])) {
                 foreach ($matches[2] as $idT => $couleur) {
-                    if(isset($matches[3][$idT])){
-                        $chaine = str_replace('#'.$matches[2][$idT].'#'.$matches[3][$idT].'#', '<span style="color:'.$matches[2][$idT].'">'.$matches[3][$idT].'</span>', $chaine);
+                    if (isset($matches[3][$idT])) {
+                        $chaine = str_replace('#' . $matches[2][$idT] . '#' . $matches[3][$idT] . '#', '<span style="color:' . $matches[2][$idT] . '">' . $matches[3][$idT] . '</span>', $chaine);
                     }
                 }
             }
@@ -300,11 +337,10 @@ class BimpDocumentation {
     static function traiteLink($str, $chaine, $mode = 'link') {
         $lienDoc = DOL_URL_ROOT . '/bimpcore/docs/test.php?name=';
         if (file_exists(static::getPathFile('doc', $str)))
-            if($mode == 'modal'){
-                $onClick = "docModal.loadAjaxContent($(this), 'loadDocumentation', {name: '".$str."'}, 'Doc : ".$str."', 'Chargement', function (result, bimpAjax) {});";
-                $chaine = '<a onclick="'.$onClick.'">' . $chaine . '</a>';
-            }
-            else
+            if ($mode == 'modal') {
+                $onClick = "docModal.loadAjaxContent($(this), 'loadDocumentation', {name: '" . $str . "'}, 'Doc : " . $str . "', 'Chargement', function (result, bimpAjax) {});";
+                $chaine = '<a onclick="' . $onClick . '">' . $chaine . '</a>';
+            } else
                 $chaine = '<a href="' . $lienDoc . $str . '">' . $chaine . '</a>';
         return $chaine;
     }
